+++ go/vt/servenv/buildinfo.go
@@ -23,22 +23,11 @@ import (
 	"strconv"
 	"time"
 
-	"vitess.io/vitess/go/vt/proto/vtrpc"
-	"vitess.io/vitess/go/vt/vterrors"
-
-	"vitess.io/vitess/go/vt/log"
-
-	"vitess.io/vitess/go/vt/sqlparser"
-
 	"vitess.io/vitess/go/stats"
+	"vitess.io/vitess/go/vt/sqlparser"
 )
 
 var (
-	// MySQLServerVersion is what Vitess will present as it's version during the connection handshake,
-	// and as the value to the @@version system variable. If nothing is provided, Vitess will report itself as
-	// a specific MySQL version with the vitess version appended to it
-	MySQLServerVersion = flag.String("mysql_server_version", "", "MySQL server version to advertise.")
-
 	buildHost             = ""
 	buildUser             = ""
 	buildTime             = ""
@@ -81,14 +70,13 @@ func (v *versionInfo) String() string {
 }
 
 func (v *versionInfo) MySQLVersion() string {
-	if *MySQLServerVersion != "" {
-		return *MySQLServerVersion
+	if *sqlparser.MySQLServerVersion != "" {
+		return *sqlparser.MySQLServerVersion
 	}
 	return "5.7.9-vitess-" + v.version
 }
 
 func init() {
-	flag.Parse()
 	t, err := time.Parse(time.UnixDate, buildTime)
 	if buildTime != "" && err != nil {
 		panic(fmt.Sprintf("Couldn't parse build timestamp %q: %v", buildTime, err))
@@ -112,13 +100,6 @@ func init() {
 		goArch:             runtime.GOARCH,
 		version:            versionName,
 	}
-	var convVersion string
-	convVersion, err = convertMySQLVersionToCommentVersion(AppVersion.MySQLVersion())
-	if err != nil {
-		log.Error(err)
-	} else {
-		sqlparser.MySQLVersion = convVersion
-	}
 	stats.NewString("BuildHost").Set(AppVersion.buildHost)
 	stats.NewString("BuildUser").Set(AppVersion.buildUser)
 	stats.NewGauge("BuildTimestamp", "build timestamp").Set(AppVersion.buildTime)
@@ -139,42 +120,4 @@ func init() {
 		fmt.Sprintf("%v", AppVersion.jenkinsBuildNumber),
 	}
 	stats.NewGaugesWithMultiLabels("BuildInformation", "build information exposed via label", buildLabels).Set(buildValues, 1)
-}
-
-// convertMySQLVersionToCommentVersion converts the MySQL version into comment version format.
-func convertMySQLVersionToCommentVersion(version string) (string, error) {
-	var res = make([]int, 3)
-	idx := 0
-	val := ""
-	for _, c := range version {
-		if c <= '9' && c >= '0' {
-			val += string(c)
-		} else if c == '.' {
-			v, err := strconv.Atoi(val)
-			if err != nil {
-				return "", err
-			}
-			val = ""
-			res[idx] = v
-			idx++
-			if idx == 3 {
-				break
-			}
-		} else {
-			break
-		}
-	}
-	if val != "" {
-		v, err := strconv.Atoi(val)
-		if err != nil {
-			return "", err
-		}
-		res[idx] = v
-		idx++
-	}
-	if idx == 0 {
-		return "", vterrors.Errorf(vtrpc.Code_INVALID_ARGUMENT, "MySQL version not correctly setup - %s.", version)
-	}
-
-	return fmt.Sprintf("%01d%02d%02d", res[0], res[1], res[2]), nil
 }
