+++ go/vt/sqlparser/parser.go
@@ -17,8 +17,10 @@ limitations under the License.
 package sqlparser
 
 import (
+	"flag"
 	"fmt"
 	"io"
+	"strconv"
 	"strings"
 	"sync"
 
@@ -27,6 +29,11 @@ import (
 
 	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 )
+
+// MySQLServerVersion is what Vitess will present as it's version during the connection handshake,
+// and as the value to the @@version system variable. If nothing is provided, Vitess will report itself as
+// a specific MySQL version with the vitess version appended to it
+var MySQLServerVersion = flag.String("mysql_server_version", "", "MySQL server version to advertise.")
 
 // parserPool is a pool for parser objects.
 var parserPool = sync.Pool{
@@ -39,7 +46,7 @@ var parserPool = sync.Pool{
 var zeroParser yyParserImpl
 
 // MySQLVersion is the version of MySQL that the parser would emulate
-var MySQLVersion string = "50709"
+var MySQLVersion string
 
 // yyParsePooled is a wrapper around yyParse that pools the parser objects. There isn't a
 // particularly good reason to use yyParse directly, since it immediately discards its parser.
@@ -97,6 +104,56 @@ func Parse2(sql string) (Statement, BindVars, error) {
 		return nil, nil, ErrEmpty
 	}
 	return tokenizer.ParseTree, tokenizer.BindVars, nil
+}
+
+func checkParserVersionFlag() {
+	if MySQLVersion == "" {
+		convVersion, err := convertMySQLVersionToCommentVersion(*MySQLServerVersion)
+		if err != nil {
+			log.Error(err)
+			MySQLVersion = "50709" // default version if nothing else is stated
+		} else {
+			MySQLVersion = convVersion
+		}
+	}
+}
+
+// convertMySQLVersionToCommentVersion converts the MySQL version into comment version format.
+func convertMySQLVersionToCommentVersion(version string) (string, error) {
+	var res = make([]int, 3)
+	idx := 0
+	val := ""
+	for _, c := range version {
+		if c <= '9' && c >= '0' {
+			val += string(c)
+		} else if c == '.' {
+			v, err := strconv.Atoi(val)
+			if err != nil {
+				return "", err
+			}
+			val = ""
+			res[idx] = v
+			idx++
+			if idx == 3 {
+				break
+			}
+		} else {
+			break
+		}
+	}
+	if val != "" {
+		v, err := strconv.Atoi(val)
+		if err != nil {
+			return "", err
+		}
+		res[idx] = v
+		idx++
+	}
+	if idx == 0 {
+		return "", vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "MySQL version not correctly setup - %s.", version)
+	}
+
+	return fmt.Sprintf("%01d%02d%02d", res[0], res[1], res[2]), nil
 }
 
 // Parse behaves like Parse2 but does not return a set of bind variables
